Pathfinding 1
In this demo, we show a more complicated map and a single agent has
to navigate it using the A* pathfinding algorithm. The current version
uses 8-neighborhood and euclidean distance as an heuristic, however,
4-neighborhood and/or manhattan distances are also implemented. The dark
blue squares represent explored tiles, the light-blue ones are the
shortest path, and the purple ones are the actual waypoints
(optimization).